+++ src/dotnet/NugetProjectProperties.ts
@@ -15,9 +15,6 @@ import {
 } from '../utils/reflection.ts';
 import { MSBuildProjectProperties, type Class_MSBPP } from './MSBuildProjectProperties.ts';
 
-const Type_RepoType: type<'' | 'git' | 'tfs'> = type('""|"git"|"tfs"');
-const Type_SymbolPackageFormat: type<'symbols.nupkg' | 'snupkg'> = type('"symbols.nupkg" | "snupkg"');
-
 /**
  * A readonly record of a .csproj or .fsproj with NuGet configuration properties in
  * addition to those specified by {@link MSBuildProjectProperties}. This record
@@ -28,6 +25,7 @@ export class NugetProjectProperties extends MSBuildProjectProperties {
   // #region private
   private _isPackable: BooleanString | undefined;
   private _suppressDependenciesWhenPacking: BooleanString | undefined;
+  private _generatePackageOnBuild: BooleanString | undefined;
   private _packageVersion: string | undefined;
   private _packageId: string | undefined;
   private _packageDescription: string | undefined;
@@ -162,6 +160,18 @@ export class NugetProjectProperties extends MSBuildProjectProperties {
    */
   get SuppressDependenciesWhenPacking(): BooleanString {
     return this._suppressDependenciesWhenPacking ??= 'false';
+  }
+
+  /**
+   * @returns "true" or "false". If "true", Pack will run after every Build. To
+   * prevent recursion, Build will no longer auto-run before Pack.
+   *
+   * WARNING: Pack may fail if there are no Build artifacts!
+   *
+   * Default: "false"
+   */
+  get GeneratePackageOnBuild(): BooleanString {
+    return this._generatePackageOnBuild ??= 'false';
   }
 
   /**
@@ -561,11 +571,6 @@ export class NugetProjectProperties extends MSBuildProjectProperties {
   }
 }
 
-export type Class_NPP = ClassLike<
-  typeof NugetProjectProperties
-  & WithPrototype<Class_MSBPP>
->;
-
 // eslint-disable-next-line @typescript-eslint/no-extraneous-class
 export class NPPGetterNames {
   private static _prototypeChain: PrototypeChainOfClass<Class_NPP> | undefined;
@@ -599,5 +604,12 @@ export class NPPGetterNames {
     );
   }
 }
+
+export type Class_NPP = ClassLike<
+  typeof NugetProjectProperties
+  & WithPrototype<Class_MSBPP>
+>;
+const Type_RepoType: type<'' | 'git' | 'tfs'> = type('""|"git"|"tfs"');
+const Type_SymbolPackageFormat: type<'symbols.nupkg' | 'snupkg'> = type('"symbols.nupkg" | "snupkg"');
 
 /** @module NugetProjectProperties */